Moves template.ts
