+++ packages/icon/scripts/template.ts
